project/requirements.txt
celery==4.4.1
Django==3.1.1
redis==3.5.3
djangorestframework==3.12.1
djoser==2.0.5
psycopg2-binary==2.8.6